Implement kernel-level support for futex objects
All of the APIs have been defined and implemented for accessing futex objects for arbitrary memory locations. The only part missing now is actually implementing the already defined system call APIs.
The most difficult part is integrating the thing into the datapart-split mechanims, but that also ended up working pretty well.

+++ kos/src/kernel/core/memory/vm/vm-datablock-read-write-base.c.inl
@@ -374,7 +374,7 @@ KCALL vm_datablock_vio_write
 	}
 #endif /* CONFIG_VIO_HAS_QWORD */
 	if (num_bytes & 2) {
-		DO_VIO(u32, w, 16);
+		DO_VIO(u16, w, 16);
 		buf = (byte_t *)buf + 2;
 		src_offset += 2;
 	}
